+++ compile_linux_kernel.sh
@@ -14,9 +14,9 @@ tput clear
 # Source whiptail messages
 . ./messages
 
+chk_version
+
 # Set overlap variables						
-#DEPENDENCIES+="build-essential dkms gnupg libncurses5 libncurses5-dev libnotify-bin \
-#				libssl-dev pkg-config qt5-qmake sudo time "
 DEPENDENCIES+="build-essential gnupg libnotify-bin libssl-dev pkg-config \
 				sudo time "