+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.0.4 (Jun 4, 2016)
 NEW FEATURES
-- none
+- Add version check
 
 IMPROVEMENTS
 - Add `post-commit` created `version` file to aid user in determining current status against repo